+++ app/javascript/legacy/components/nonprofit-branding.js
@@ -1,5 +1,5 @@
 // License: LGPL-3.0-or-later
-import nonprofitBranding from '../../../javascripts/src/lib/nonprofitBranding.ts';
+import nonprofitBranding from '../../legacy_react/src/lib/nonprofitBranding.ts';
 
 module.exports = nonprofitBranding(app.nonprofit.brand_color)
