+++ app/javascript/legacy/recurring_donations/edit/custom-nonprofit-branding.es6
@@ -1,4 +1,4 @@
 // License: LGPL-3.0-or-later
-import nonprofitBranding from '../../../../javascripts/src/lib/nonprofitBranding.ts';
+import nonprofitBranding from '../../../legacy_react/src/lib/nonprofitBranding';
 
 module.exports = nonprofitBranding
