+++ app/javascript/packs/application.js
@@ -1,4 +1,3 @@
 import "core-js/stable";
 import "regenerator-runtime/runtime";
 
-require('../legacy/recurring_donations/edit/page')
