+++ app/javascript/packs/edit_payment_pane.js
@@ -1,0 +1,4 @@
+// License: LGPL-3.0-or-later
+// require a root component here. This will be treated as the root of a webpack package
+require('bootstrap-loader');
+require('../legacy_react/app/edit_payment_pane')
