+++ app/javascript/packs/page__bank_accounts__confirm.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/bank_accounts/confirm/page.js')
