+++ app/javascript/packs/page__campaigns__index.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/campaigns/index/page.js')
