+++ app/javascript/packs/page__campaigns__peer_to_peer.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/campaigns/peer_to_peer/page.js')
