+++ app/javascript/packs/page__campaigns__supporters__index.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/campaigns/supporters/index/page.js')
