+++ app/javascript/packs/page__events__show.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/events/show/page.js')
