+++ app/javascript/packs/page__events__stats.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/events/stats/page.js')
