+++ app/javascript/packs/page__nonprofits__button.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/button/page.js')
