+++ app/javascript/packs/page__nonprofits__dashboard.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/dashboard/page.js')
