+++ app/javascript/packs/page__nonprofits__donate.js
@@ -1,0 +1,3 @@
+
+require('../legacy/page.js')
+require('../legacy/nonprofits/donate/page.js')
