+++ app/javascript/packs/page__nonprofits__payouts__index.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/payouts/index/page.js')
