+++ app/javascript/packs/page__nonprofits__recurring_donations__index.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/recurring_donations/index/page.js')
