+++ app/javascript/packs/page__nonprofits__show.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/show/page.js')
