+++ app/javascript/packs/page__nonprofits__supporter_form.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/nonprofits/supporter_form/page.js')
