+++ app/javascript/packs/page__nonprofits__supporters__index.js
@@ -1,0 +1,2 @@
+require('../legacy/page.js')
+require('../legacy/nonprofits/supporters/index/page.js')
