+++ app/javascript/packs/page__profiles__fundraisers.js
@@ -1,0 +1,2 @@
+
+require('../legacy/page.js')
