+++ app/javascript/packs/page__recurring_donations__edit.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/recurring_donations/edit/page.js')
