+++ app/javascript/packs/page__settings__index.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/settings/index/page.js')
