+++ app/javascript/packs/page__super-admin.js
@@ -1,0 +1,3 @@
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/super-admin/page.js')
