+++ app/javascript/packs/page__tickets__index.js
@@ -1,0 +1,4 @@
+
+require('../../../client/css/global/page.css')
+require('../legacy/page.js')
+require('../legacy/tickets/index/page.js')
